chore: ruff lint stragglers from Stage 6
* `f"...".encode("utf-8")` -> `.encode()` (UP012). UTF-8 is the
  default and the explicit argument is just noise.
* `int(round(...))` -> `round(...)` (RUF046). `round(int|float)`
  returns int already; the cast is redundant.

No behavior change.

+++ tests/test_retrieve_perf.py
@@ -38,7 +38,7 @@ COLD_REBUILD_BUDGET_MS = 1500.0
 
 def _percentile(values: list[float], p: float) -> float:
     s = sorted(values)
-    idx = int(round((p / 100.0) * (len(s) - 1)))
+    idx = round((p / 100.0) * (len(s) - 1))
     return s[idx]
